% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_VirginiaC19.R
\name{refresh_VirginiaC19}
\alias{refresh_VirginiaC19}
\title{Refresh the Virginia COVID-19 Data.}
\source{
\href{https://github.com/debusklaneml/VirginiaC19}{VirginiaC19}

\href{https://github.com/Covid19R/documentation}{covid19R documentation}
}
\usage{
refresh_VirginiaC19(verbose = TRUE)
}
\arguments{
\item{verbose}{Should messages be displayed?}
}
\value{
A tibble object
\itemize{
\item date - The date in YYYY-MM-DD form
\item location - The name of the location as provided by the data source.
\item location_type - The type of location using the covid19R controlled vocabulary.
\item location_code - A standardized location code using a national or international standard.
\item location_code_type The type of standardized location code being used according to the covid19R controlled vocabulary.
\item data_type - the type of data in that given row using the covid19R controlled vocabulary. Daily totals are found here.
\item value - number of cases of each data type
}
}
\description{
Updates the Virginia COVID-19 cases.
}
\details{
This package scrapes the state of Virginia reported cases.
}
\examples{
\dontrun{
refresh_VirginiaC19()
}

}
