% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VirginiaC19}
\alias{VirginiaC19}
\title{VirginiaC19 cases data.}
\format{
A tibble object
}
\source{
\href{https://www.vdh.virginia.gov/coronavirus/}{website}
}
\usage{
VirginiaC19
}
\description{
daily summary of Virginia Coronavirus (COVID-19) cases.
}
\details{
This package includes tidy data of Virginia COVID-19 cases.

To get the latest data, use \code{\link{refresh_VirginiaC19}}
}
\examples{

data(VirginiaC19)

# Show us something cool
}
\keyword{COVID19}
\keyword{Virginia}
\keyword{coronavirus}
\keyword{datasets}
